requirements.txt
Django==4.2.0
django-admin-sortable2==2.1.10
django-tinymce==4.0.0
Pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
